+++ app/Http/Controllers/SkillController.php
@@ -38,11 +38,13 @@ final class SkillController extends Controller
          */
         $validated = $request->validate([
             'name' => ['required', 'string', 'min:3', 'max:255', UniqueTranslationRule::for('tags')->where('type', TagType::SKILL->value)],
+            'is_regular' => ['required', 'boolean'],
         ]);
 
         Tag::create([
             'name' => $validated['name'],
             'type' => TagType::SKILL->value,
+            'is_regular' => $validated['is_regular'],
         ]);
 
         return to_route('skills.index')->with(FlashMessageKey::SUCCESS->value, __('Skill created successfully.'));
@@ -55,8 +57,12 @@ final class SkillController extends Controller
     {
         $tag = Tag::findOrFail($id);
 
+        if ($tag->is_regular) {
+            return to_route('skills.index')->with(FlashMessageKey::ERROR->value, __('Cannot update a regular skill.'));
+        }
+
         /**
-         * @var array{name:string}
+         * @var array{name:string,is_regular:bool}
          */
         $validated = $request->validate([
             'name' => ['required', 'string', 'min:3', 'max:255',
@@ -64,10 +70,12 @@ final class SkillController extends Controller
                     ->ignore($tag->id)
                     ->where('type', TagType::SKILL->value),
             ],
+            'is_regular' => ['required', 'boolean'],
         ]);
 
         $tag->update([
             'name' => $validated['name'],
+            'is_regular' => $validated['is_regular'],
         ]);
 
         return to_route('skills.index')->with(FlashMessageKey::SUCCESS->value, __('Skill updated successfully.'));
@@ -79,6 +87,9 @@ final class SkillController extends Controller
     public function destroy(string $id): RedirectResponse
     {
         $tag = Tag::findOrFail($id);
+        if ($tag->is_regular) {
+            return to_route('skills.index')->with(FlashMessageKey::ERROR->value, __('Cannot delete a regular skill.'));
+        }
         $tag->delete();
 
         return to_route('skills.index')->with(FlashMessageKey::SUCCESS->value, __('Skill deleted successfully.'));
